Login Cliente Listo

+++ REST_API_GymTEC/Models/Client.cs
@@ -11,7 +11,7 @@
         public string direccion { get; set; } = string.Empty;
         public string email { get; set; } = string.Empty;
         public string password { get; set; } = string.Empty;
-        public float IMC { get; set; }
+        public double IMC { get; set; }
         public int edad { get; set; }
 
     }
